REsolved conflicts, please push this as it is
