+++ Components/Home Page/categories/Categories.jsx
@@ -1,6 +1,6 @@
 import "@fortawesome/fontawesome-free/css/all.min.css";
 import "@fortawesome/fontawesome-free/js/all.js";
-import styles from "./categories.module.css";
+import styles from "./styles/categories.module.css";
 
 let category = [
   { icon: "fa-solid fa-chart-line", name: "Sales" },
@@ -15,7 +15,7 @@ let category = [
 function Categories() {
   return (
     <section className={styles.parent}>
-      <h2 className={styles["header"]}>Categories</h2>
+      <h2 className="header-text">Categories</h2>
       <ul>
         {category.map((category) => (
           <li key={category.name} className={`${category.icon} ${styles.li} `}>
