+++ Components/Home Page/featured courses/FeaturedCourses.jsx
@@ -35,11 +35,11 @@ const courses = [
     price: "24",
   },
 ];
-
+// const courses = fetch()
 const FeaturedCourses = () => {
   return (
     <section className={styles["features-container"]}>
-      <h2 className={styles["header"]}>Featured Courses</h2>
+      <h2 className="header-text">Featured Courses</h2>
       <ul className={`${styles["cards-container"]} mflex`}>
         <DisplayCourses AllCourses={courses}></DisplayCourses>
       </ul>
