+++ Components/Home Page/featured courses/styles/featuredCourseElement.module.css
