--- app/Components/Home Page/categories/styles/categories.module.css
@@ -1,27 +1,0 @@
-.category {
-  text-align: center;
-  display: inline-block;
-  padding-top: 10px;
-  font-size: 15px;
-  background-color: white;
-  border-radius: 20px;
-  width: 140px;
-  height: 70px;
-  margin: 5px;
-  padding-top: 20px;
-  padding-bottom: 30px;
-}
-
-.category:hover {
-  scale: 1.15;
-  border: 2px solid rgb(46 141 255);
-  background-color: rgb(232 239 246);
-  color: rgb(46 141 255);
-}
-
-.text {
-  font-size: 15px;
-  font-weight: 500;
-  font-family: 'Courier New', Courier, monospace;
-  margin-top: 10px;
-}
