--- app/Components/Home Page/my learning/styles/myLearning.module.css
@@ -1,19 +1,0 @@
-.header {
-  font-size: 3rem;
-  font-weight: 600;
-  color: #303030;
-  margin-bottom: 2rem;
-}
-
-.myLearning-container {
-}
-
-.cards-container {
-  gap: 1rem;
-}
-
-@media (max-width: 800px) {
-  .cards-container {
-    flex-direction: column;
-  }
-}
