+++ app/home/page.jsx
@@ -1,8 +1,7 @@
-import Categories from "../Components/Home Page/categories/Categories";
-import FeaturedCourses from "../Components/Home Page/featured courses/FeaturedCourses";
-import MyLearning from "../Components/Home Page/my learning/MyLearning";
+import Categories from "@/Components/Home Page/categories/Categories";
+import FeaturedCourses from "@/Components/Home Page/featured courses/FeaturedCourses";
+import MyLearning from "@/Components/Home Page/my learning/MyLearning";
 import styles from "./homePage.module.css";
-
 const Home = (props) => {
   return (
     <main className={styles["home-container"]}>
