+++ app/search/searchPage.module.css
@@ -1,0 +1,14 @@
+.home-container {
+  margin: auto 12rem auto 4rem;
+}
+@media (max-width: 950px) {
+  .home-container {
+    margin-right: 4rem;
+  }
+}
+@media (max-width: 480px) {
+  .home-container {
+    margin-right: 1rem;
+    margin-left: 1rem;
+  }
+}
